feat: add new keys in qc_organization and improve flexibility
The dict qc_organization will receive additional keys as features are added. The code and documentation have been modified to minimize the changes needed when keys are added. At this stage only ampyl.py needs to be modified in three places: the doc-string of FiniteVolumeSetup, the default __init__ of the same class and the setter method for qc_organization.