fix: boost momentum cutoff back to the lab frame for n1n2n3 sets
The non-interacting states are assembled based on Emax and Lmax. This requires choosing maximum momnetum components to loop over. For three-particle states this was done by setting the spectator to be at rest. However the code mistakenly delivered the maximum possible back to back momentum in the overall centre-of-mass frame. This is now boosted back to the finite-volume frame.