feat: add qc_impl option g_uses_prep_mat
The QCIndexSpace contains all information about the matrix space of the quantization that is independent of the exact values of volume and energy. Previously, the default for building the G-matrix was to promote vectors of spectator momenta to matrices everytime G is calculated. This update adds a feature that prepares more matrices in advance, but only for the total-momentum zero case. Speed-up is quite modest/dissapointing.