feat: add major interpolation support for G
The evaluation of the G matrix can be a significant fraction of the computation time in computing the quantization condition. The cost partly arises from the large matrices that are evaluated and then projected down to smaller matrices corresponding to a given irrep. The added methods support interpolating the entries of the projected matrices (so the large matrices are avoided in sampling the quantization condition). This also requires removing poles before interpolating and then re-introducing these in the evaluation stage. Testing is urgently needed.