feat: include reduce_size flag for nonzero nP interp
A major challege for interpolation is that the size of matrices changes with E and L. The handling of this is solved for the nP=0 case, but in the moving frame the growth of matrices with E and L is more complicated. In particular, it is unclear how to match up matrix elements on either side of a shell opening. This feature simply keeps the largest matrices, set by Emax and Lmax, with many zeros for lower E and L.