feat: improve pole insertion and add non-int multiplicites
Poles have to be inserted whenever get_value is called on an Interpolable with an interpolation option set to True. This was previously done with a python for loop and is now done using prepared numpy arrays for poles and pole matrix textures. Numpy sum and prod can be used for all steps. Speed up is only noticeable when many poles contribute. Additionally, non-interacting multiplicties are now contained within QCIndexSpace().nonint_multiplicities.