feat: support moving frame interpolators
A number of changes were needed to get interpolation working for non-zero total momentum. For now we are assuming one index space throughout, based on Emax and Lmax. We also insist that interpolation grids all have the same range with zeros included when needed. This allows one to change loop order when searching for poles: a major speed up.