test: adjust trivial tests on channel construction
Two tests were failing. The first was feeding isospin particles into a non-isospin channel so the failure looks correct and is fixed. The second was based on the dimensions of a qcis. This was adjusted to pass with the old version left as a comment in test_load_ni_data. This should be insprected more carefully.