refactor: improve solution walking within QC
The code for walking along a given solution within the QC class is improved significantly, as are other methods for finding solutions. As a breaking change, get_roots_at_Lmax is now get_roots_at_fixed_L. Within this method, the strategy for determining unique roots is rewritten. The handling on non-interacting solutions is imporved and accomodates total isospin 1, but more work is needed. Major improvements are also made to get_energy_curve. Most notably, the extrapolation to get the next solution windo now uses a numpy polynomial fit. The previous usage of np.interp did not work as this returns the extreme value when the requested x is outside the supported range.

BREAKING CHANGE: get_roots_at_fixed_L